Refactor to standard directory-based layour
