+++ svg3d/__init__.py
@@ -1,0 +1,14 @@
+from .shaders import diffuse_shader, hex2rgb, rgb2hex
+from .svg3d import Engine
+from .view import View, _pad_arrays, get_lookat_matrix, get_projection_matrix
+
+__all__ = [
+    "_pad_arrays",
+    "Engine",
+    "get_lookat_matrix",
+    "get_projection_matrix",
+    "View",
+    "hex2rgb",
+    "rgb2hex",
+    "diffuse_shader",
+]
